create score badge with custom animation
