+++ memorize/View/Components/Badges/ScoreBadge.swift
@@ -16,6 +16,7 @@ struct ScoreBadge: View {
                 .contentTransition(.numericText())
                 .animation(.easeInOut(duration: 0.25), value: score)
         }
+        .bold()
         .lineLimit(1)
         .fixedSize(horizontal: true, vertical: false)
         .layoutPriority(1)
@@ -34,28 +35,3 @@ struct ScoreBadge: View {
         }
     }
 }
-
-
-struct ShakeEffect: GeometryEffect {
-    enum Axis { case horizontal, vertical }
-    
-    var amount: CGFloat = 6
-    var shakes: CGFloat = 0
-    var axis: Axis = .horizontal
-    
-    var animatableData: CGFloat {
-        get { shakes }
-        set { shakes = newValue }
-    }
-    
-    func effectValue(size: CGSize) -> ProjectionTransform {
-        let offset = amount * sin(shakes * .pi * 2)
-        switch axis {
-        case .horizontal:
-            return ProjectionTransform(.init(translationX: offset, y: 0))
-        case .vertical:
-            return ProjectionTransform(.init(translationX: 0, y: offset))
-        }
-    }
-}
-
